// ui/src/models/applicationset.ts
export interface ApplicationSet {
    apiVersion?: string;
    kind?: string;
    metadata: models.ObjectMeta;
    spec: ApplicationSetSpec;
    status?: ApplicationSetStatus;
}

export interface ApplicationSetSpec {
    goTemplate?: boolean;
    generators: ApplicationSetGenerator[];
    template: ApplicationSetTemplate;
    syncPolicy?: ApplicationSetSyncPolicy;
}

export interface ApplicationSetSyncPolicy {
    preserveResourcesOnDeletion?: boolean;
}

export interface ApplicationSetTemplate {
    spec: { [key: string]: any};
}

export interface ApplicationSetTemplateMeta {
    name?: string;
    namespace?: string;
    labels?: { [key: string]: string};
    annotations?: { [key: string]: string};
    finalizers?: string[];
}

export interface ApplicationSetGenerator {
    list?: ListGenerator;
    clusters?: ClusterGenerator;
    git?: GitGenerator;
    scmProvider?: SCMProviderGenerator;
    clusterDecisionResource?: DuckTypeGenerator;
    pullRequest?: PullRequestGenerator;
    matrix?: MatrixGenerator;
    merge?: MergeGenerator;
    selector?: { [key: string]: string};
}
// ...
export interface ApplicationSetStatus {
    conditions?: ApplicationSetCondition[];
}

export type ApplicationSetConditionType = 'ErrorOccurred' | 'ParametersGenerated' | 'ResourcesUpToDate';

export type ApplicationSetConditionStatus = 'True' | 'False' | 'Unknown'

export interface ApplicationSetCondition {
    type: ApplicationSetConditionType;
    message: string;
    lastTransitionTime?: string;
    status: ApplicationSetConditionStatus;
    reason: string;
}